refactor: system/images 로 분리
